Moved state machine test to its own package
