+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/statemachine/StateOne.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.statemachine;
 
 import com.qualcomm.robotcore.util.ElapsedTime;
