+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/statemachine/StateTwo.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.statemachine;
 
 import team.techtigers.base.statemachine.State;
